refactor: 고정비 페이지 SSR 구조 초안 구성 #106
