+++ fe/src/app/(app)/fixed-costs/page.tsx
@@ -1,5 +1,32 @@
-import FixedCostsPage from '@/components/fixed-costs/FixedCostsPage';
+import FixedCostsClient from '@/components/fixed-costs/FixedCostsClient';
+import FixedCostsList from '@/components/fixed-costs/FixedCostsList';
+import FixedCostsListSkeleton from '@/components/fixed-costs/FixedCostsListSkeleton';
+import { fetchFixedRules } from '@/services/fixed-costs/fixed-costs';
+import { requireUserServer } from '@/utils/supabase/requireUserServer';
+import { Suspense } from 'react';
 
-export default function Page() {
-  return <FixedCostsPage />;
+export default function FixedCostsPage() {
+  return (
+    <div className="flex min-h-screen w-full flex-col">
+      <div className="flex w-full flex-col items-center space-y-6 p-4 md:p-6 lg:p-8">
+        <div className="w-full max-w-xl space-y-6">
+          <header>
+            <h1 className="text-xl font-semibold">고정비 관리</h1>
+          </header>
+        </div>
+      </div>
+      <FixedCostsClient />
+
+      <Suspense fallback={<FixedCostsListSkeleton />}>
+        <FixedCostsListWrapper />
+      </Suspense>
+    </div>
+  );
+}
+
+async function FixedCostsListWrapper() {
+  const { user } = await requireUserServer();
+  const items = await fetchFixedRules(user.id);
+
+  return <FixedCostsList items={items} isLoading={false} onEdit={() => {}} />;
 }
